move github client tests to separate module
