+++ tests/integration/test_target_branch_protection.py
@@ -4,14 +4,11 @@
 """Tests for the target_branch_protection function."""
 
 import itertools
-import typing
 from uuid import uuid4
 
 import pytest
 
 from repo_policy_compliance import Result, target_branch_protection
-from repo_policy_compliance.exceptions import GithubClientError, InputError
-from repo_policy_compliance.github_client import GITHUB_TOKEN_ENV_NAME
 
 from .. import assert_
 from .types_ import BranchWithProtection
@@ -102,40 +99,3 @@ def test_pass(
 
     assert report.reason is None
     assert report.result == Result.PASS
-
-
-@pytest.mark.parametrize(
-    "github_token_value, exception_context, exception_message",
-    [
-        pytest.param("", pytest.raises(InputError), "was not provided", id="github_token empty"),
-        pytest.param(
-            None, pytest.raises(GithubClientError), "Bad Credential error", id="github_token none"
-        ),
-        pytest.param(
-            "abcd",
-            pytest.raises(GithubClientError),
-            "Bad Credential error",
-            id="github_token wrong",
-        ),
-    ],
-)
-def test_github_token(
-    github_token_value: str | None,
-    exception_context: typing.ContextManager,
-    exception_message: str,
-    github_repository_name: str,
-    monkeypatch: pytest.MonkeyPatch,
-):
-    """
-    arrange: A github repository name and a missing github token.
-    act: when the github client is injected to target_branch_protection.
-    assert: An expected error is raised with a specific error message.
-    """
-    monkeypatch.delenv(GITHUB_TOKEN_ENV_NAME)
-    monkeypatch.setenv(GITHUB_TOKEN_ENV_NAME, str(github_token_value))
-    with exception_context as error:
-        # The github_client is injected
-        target_branch_protection(  # pylint: disable=no-value-for-parameter
-            repository_name=github_repository_name, branch_name="arbitrary"
-        )
-    assert_.substrings_in_string([GITHUB_TOKEN_ENV_NAME, exception_message], str(error.value))
